move NamespaceName to own module
